+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 7.0.2 (Unreleased)
 
+**New Features**
+
+* Added a `body_as_str` method on `ServiceBusMessage` which returns the content of the message as a string.
 
 ## 7.0.1 (2021-01-12)